value.php to prevent web server logs overfill
Add a sleep 60 to prevent the web server log overfilling with errors when a sensor is offline or inaccessible. Otherwise, the php script will continuously generate errors in the do while loop and every loop is recorded by the Apache error log. The log will fill up all the space of the SD card within a couple of days.